+++ libft/ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: mmoulati <mmoulati@student.1337.ma>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/10 12:19:08 by mmoulati          #+#    #+#             */
-/*   Updated: 2024/11/10 12:19:08 by mmoulati         ###   ########.fr       */
+/*   Updated: 2024/11/10 13:05:15 by mmoulati         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i])
